requirements.txt
asgiref==3.7.2
astroid==2.15.5
black==23.3.0
click==8.1.3
dill==0.3.6
dj-rest-auth==4.0.1
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
isort==5.12.0
lazy-object-proxy==1.9.0
Markdown==3.4.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.3
psycopg2-binary==2.9.6
PyJWT==2.7.0
pylint==2.17.4
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.3
wrapt==1.15.0
